fix: overflow property set to 'clip' (instead of 'hidden')
this will fix the baseline issue wich caused the content to jump
when dynamically setting 'overflow: hidden'
instead of an inline css rule, a css class is now being used to set
the overflow property. @supports was also added to ensure 'clip' will
only be used on supported devices. if not supported, will use 'hidden' instead